File with useful links I keep for my students
This is most of the file I share with my students that contains a bunch of links, e.g., intro to regular expressions, useful commands in Unix, etc. Mark had asked me to add the file to the course repo. Let me know if you want me to format that in a different way